backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "ddgs==9.6.1",
    "fastapi[all]>=0.118.0",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.34",
    "openai>=2.1.0",
    "pydantic>=2.11.9",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
]
